ml/training/requirements.txt
pandas
scikit-learn
joblib
google-cloud-bigquery